some refactoring of util
